+++ CHANGELOG.md
@@ -5,6 +5,61 @@
 
 > All notable changes to this project will be documented in this file
 
+## [1.9.0-beta.2](https://github.com/open-sauced/ai/compare/v1.9.0-beta.1...v1.9.0-beta.2) (2023-06-21)
+
+
+### 🐛 Bug Fixes
+
+* separate vite and vitest configs ([#193](https://github.com/open-sauced/ai/issues/193)) ([7a980c8](https://github.com/open-sauced/ai/commit/7a980c82d1056cd1f798c39e51ba9fbc4d205ad7))
+
+## [1.9.0-beta.1](https://github.com/open-sauced/ai/compare/v1.8.1-beta.1...v1.9.0-beta.1) (2023-06-21)
+
+
+### 🍕 Features
+
+* post to linkedin projects ([#177](https://github.com/open-sauced/ai/issues/177)) ([a20bff4](https://github.com/open-sauced/ai/commit/a20bff4aaa762b40fb5d80da8a02ee76271f3ac4))
+
+## [1.8.1-beta.1](https://github.com/open-sauced/ai/compare/v1.8.0...v1.8.1-beta.1) (2023-06-20)
+
+
+### ✅ Tests
+
+* Tests for colorPreference and getUserPRVelocity ([#190](https://github.com/open-sauced/ai/issues/190)) ([6b9a5c3](https://github.com/open-sauced/ai/commit/6b9a5c320e07b10e4222d7210f8f8b7917e6abe7))
+
+## [1.8.0](https://github.com/open-sauced/ai/compare/v1.7.0...v1.8.0) (2023-06-18)
+
+
+### 🧑‍💻 Code Refactoring
+
+* move ai config to settings ([#183](https://github.com/open-sauced/ai/issues/183)) ([c31bd61](https://github.com/open-sauced/ai/commit/c31bd61d77878465c359455294865007ff07e7c0))
+* remove link to dashboard ([#185](https://github.com/open-sauced/ai/issues/185)) ([25b0017](https://github.com/open-sauced/ai/commit/25b0017ea538e20934ce7b0f751e7903a5d5276b))
+
+
+### 🍕 Features
+
+* show highlight reactions ([#186](https://github.com/open-sauced/ai/issues/186)) ([6707b0b](https://github.com/open-sauced/ai/commit/6707b0bf3abe67f9b9a26821c8ae4c8eeafb6d07))
+
+## [1.8.0-beta.1](https://github.com/open-sauced/ai/compare/v1.7.1-beta.2...v1.8.0-beta.1) (2023-06-17)
+
+
+### 🍕 Features
+
+* show highlight reactions ([#186](https://github.com/open-sauced/ai/issues/186)) ([6707b0b](https://github.com/open-sauced/ai/commit/6707b0bf3abe67f9b9a26821c8ae4c8eeafb6d07))
+
+## [1.7.1-beta.2](https://github.com/open-sauced/ai/compare/v1.7.1-beta.1...v1.7.1-beta.2) (2023-06-16)
+
+
+### 🧑‍💻 Code Refactoring
+
+* remove link to dashboard ([#185](https://github.com/open-sauced/ai/issues/185)) ([25b0017](https://github.com/open-sauced/ai/commit/25b0017ea538e20934ce7b0f751e7903a5d5276b))
+
+## [1.7.1-beta.1](https://github.com/open-sauced/ai/compare/v1.7.0...v1.7.1-beta.1) (2023-06-16)
+
+
+### 🧑‍💻 Code Refactoring
+
+* move ai config to settings ([#183](https://github.com/open-sauced/ai/issues/183)) ([c31bd61](https://github.com/open-sauced/ai/commit/c31bd61d77878465c359455294865007ff07e7c0))
+
 ## [1.7.0](https://github.com/open-sauced/ai/compare/v1.6.0...v1.7.0) (2023-06-15)